phase-d(todo): bundler/OutputFile.rs: port Value::to_bun_string fallback arm to bun_core::todo_panic

+++ src/ini/lib.rs
@@ -1387,7 +1387,7 @@ pub fn load_npmrc(
         if let Some(str_) = query.expr.as_utf8_string_literal() {
             install.ca = Some(bun_api::Ca::Str(Box::<[u8]>::from(str_)));
         } else if let ExprData::EArray(arr) = &query.expr.data {
-            let mut list: Vec<Box<[u8]>> = Vec::with_capacity(arr.items.len());
+            let mut list: Vec<Box<[u8]>> = Vec::with_capacity(arr.items.len() as usize);
             for item in arr.items.slice() {
                 if let Some(s) = item.as_string_cloned(bump)? {
                     list.push(Box::<[u8]>::from(s));
